Cleaning and Improving CodeBase
