+++ Autotype/core.py
@@ -5,26 +5,30 @@ from dataclasses import dataclass
 
 @dataclass
 class Autotype:
-    default_delay: int = 3
-    default_code: str = """
-    class Complex:
-        def __init__(self, x, y):
-            self.x = x
-            self.y = y
     """
 
-    # def __init__(self, path: str = None, delay: int = None, code: str = None):
-    #     self.path = path
-    #     self.delay = delay
-    #     self.code = code
-
-    def type(self, path: str = None, delay: int = None, code: str = None):
+    """
+    default_delay: int = 3
+    default_code: str = """
+class Main:
+    def __init__(self, name, age):
+        self.name = name
+        self.age = age
+        
+    def __str__(self):
+        return f"The name is {self.name} and the age is {self.age{"
+"""
+    @classmethod
+    def type(cls, path: str = None, delay: int = None, code: str = None):
+        """
+        
+        """
         if not delay:
-            delay = self.default_delay
+            delay = cls.default_delay
 
         if delay < 0:
             print()
-            delay = abs(self.default_delay)
+            delay = abs(cls.default_delay)
 
         sleep(delay)
 
@@ -34,7 +38,7 @@ class Autotype:
         elif code:
             pass
         else:
-            code = self.default_code
+            code = cls.default_code
 
         keyboard = Controller()
         for line in code.split("\n"):
