--- Simulator/__init__.py
@@ -1,1 +1,0 @@
-from .simulate_keyboard import Type
